index.html
If you are looking for a Responsive Table with the best quality design then you are in the right place, This table is for a design for user-interface to show table data with given table heading’s. It looks like very simple but design with css3 properties with HTML tag’s with fully responsive. In this table used the image for the background and top heading with google fonts.